+++ .gitignore
@@ -1,3 +1,5 @@
 bin/
 obj/
 .vs/
+.idea/
+*.user
