--- src/Floatingman.Common.Usage/Floatingman.Common.Usage.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="..\Floatingman.Common\Floatingman.Common.csproj" />
-  </ItemGroup>
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-<TargetFramework>net6.0</TargetFramework>
-  </PropertyGroup>
-
-</Project>
