+++ src/Floatingman.Common/EnvironmentVariables.cs
@@ -1,17 +1,14 @@
-using System;
+namespace Floatingman.Common;
 
-namespace Floatingman.Common
+public static class EnvironmentVariables
 {
-    public static class EnvironmentVariables
+    public static string GetEnvironmentValue(string key)
     {
-        public static string GetEnvironmentValue(string key)
-        {
-            return Environment.GetEnvironmentVariable(key, EnvironmentVariableTarget.Process);
-        }
-
-        //public static string GetEnvironmentVariable(string name)
-        //{
-        //    return name + ": " + GetEnvironmentValue(name);
-        //}
+        return Environment.GetEnvironmentVariable(key, EnvironmentVariableTarget.Process);
     }
+
+    //public static string GetEnvironmentVariable(string name)
+    //{
+    //    return name + ": " + GetEnvironmentValue(name);
+    //}
 }
