+++ src/Floatingman.Common/Extensions/BooleanExtensions.cs
@@ -1,15 +1,14 @@
-﻿namespace Floatingman.Common.Extensions
-{
-    public static class BooleanExtensions
-    {
-        public static bool IsNot(this bool value)
-        {
-            return !value;
-        }
+﻿namespace Floatingman.Common.Extensions;
 
-        public static bool Not(bool value)
-        {
-            return !value;
-        }
+public static class BooleanExtensions
+{
+    public static bool IsNot(this bool value)
+    {
+        return !value;
+    }
+
+    public static bool Not(bool value)
+    {
+        return !value;
     }
 }
