+++ src/Floatingman.Common/Functional/Functions.cs
@@ -1,16 +1,13 @@
-﻿using System;
+﻿namespace Floatingman.Common.Functional;
 
-namespace Floatingman.Common.Functional
+public static class Functions
 {
-    public static class Functions
+    public static TResult Using<TDisposable, TResult>(TDisposable disposable, Func<TDisposable, TResult> function)
+        where TDisposable : IDisposable
     {
-        public static TResult Using<TDisposable, TResult>(TDisposable disposable, Func<TDisposable, TResult> function)
-            where TDisposable : IDisposable
+        using (disposable)
         {
-            using (disposable)
-            {
-                return function(disposable);
-            }
+            return function(disposable);
         }
     }
 }
