+++ src/Floatingman.Common/Functional/Unit.cs
@@ -1,20 +1,35 @@
 using Unit = System.ValueTuple;
-using System;
 
-namespace Floatingman.Common.Functional
+namespace Floatingman.Common.Functional;
+
+// This allows Unit() to behave as expected.
+using static Functional;
+
+public static partial class Functional
 {
-    // This allows Unit() to behave as expected.
-    using static Functional;
-
-    public static partial class Functional
+    public static Unit Unit()
     {
-        public static Unit Unit() => default(Unit);
+        return default;
+    }
+}
+
+public static class UnitExtensions
+{
+    public static Func<Unit> ToFunc(this Action action)
+    {
+        return () =>
+        {
+            action();
+            return Unit();
+        };
     }
 
-    public static class UnitExtensions
+    public static Func<T, Unit> ToFunc<T>(this Action<T> action)
     {
-        public static Func<Unit> ToFunc(this Action action) => () => { action(); return Unit(); };
-        public static Func<T, Unit> ToFunc<T>(this Action<T> action) => (t) => { action(t); return Unit(); };
-    };
-
+        return t =>
+        {
+            action(t);
+            return Unit();
+        };
+    }
 }
