+++ tests/Floatingman.Common.xUnit/Extensions/StringExtensionsTests.cs
@@ -1,18 +1,14 @@
-using System;
-using Xunit;
-using Floatingman.Common;
-using FluentAssertions;
 using Floatingman.Common.Extensions;
+using FluentAssertions;
 
-namespace Floatingman.Common.Tests
+namespace Floatingman.Common.Tests;
+
+public class StringExtensionsTests
 {
-  public class StringExtensionsTests
-  {
-
     [Fact]
     public void DogTest()
     {
-      true.Should().BeTrue();
+        true.Should().BeTrue();
     }
 
     [Theory]
@@ -20,7 +16,6 @@ namespace Floatingman.Common.Tests
     [InlineData("a", "a")]
     public void KabobWorksAsExpected(string input, string expected)
     {
-      input.ToKabobCase().Should().Be(expected);
+        input.ToKabobCase().Should().Be(expected);
     }
-  }
 }
