+++ usage/Floatingman.Common.Usage/Floatingman.Common.Usage.csproj
@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
+    <ItemGroup>
+        <ProjectReference Include="..\..\src\Floatingman.Common\Floatingman.Common.csproj" />
+    </ItemGroup>
+
     <PropertyGroup>
+        <OutputType>Exe</OutputType>
         <TargetFramework>net7.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.3"/>
-        <PackageReference Include="YamlDotNet" Version="13.2.0"/>
-    </ItemGroup>
 
 </Project>
